+++ src/main/kotlin/auth/domain/TokenProperties.kt
@@ -1,4 +1,4 @@
-package member.config
+package auth.domain
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.stereotype.Component
