+++ src/main/kotlin/auth/domain/vo/Token.kt
@@ -1,0 +1,6 @@
+package auth.domain.vo
+
+data class Token(
+    val accessToken: AccessToken
+) {
+}
