--- src/main/kotlin/member/domain/vo/Token.kt
@@ -1,6 +1,0 @@
-package member.domain.vo
-
-data class Token(
-    val accessToken: String
-) {
-}
